Revert "new mesh extension"
This reverts commit d39c854538fd5673576ad00981241a0be4b45d2a.
